GeminiBot/requirements.txt
streamlit
google-generativeai
transformers
langchain
langchain_community
langchain_google_genai
PyPDF2
faiss-cpu
python-dotenv
